Revert "ALSA: usx2y: Fix potential NULL pointer dereference"
This reverts commit a2c6433ee5a35a8de6d563f6512a26f87835ea0f.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

The original patch was incorrect, and would leak memory if the error
path the patch added was hit.  As this is a path that is pretty
impossible to ever hit, just revert this change as it's not worth the
unwinding at this point in time.

Cc: Aditya Pakki <pakki001@umn.edu>
Reviewed-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>