Revert "isdn: mISDN: Fix potential NULL pointer dereference of kzalloc"
This reverts commit 38d22659803a033b1b66cd2624c33570c0dde77d.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

While it looks like the original change is correct, it is not, as none
of the setup actually happens, and the error value is not propagated
upwards.  So this is worse than the original code being that an atomic
operation can almost never fail, and if it did, a oops would be good
here, instead of failing to do something that the driver now claims has
happened, possibly getting state machines out of sync.  And you don't
want to mess with ISDN state machines...

Cc: Aditya Pakki <pakki001@umn.edu>
Cc: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>