Revert "dmaengine: qcom_hidma: Check for driver register failure"
This reverts commit a474b3f0428d6b02a538aa10b3c3b722751cb382.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

The original change is NOT correct, as it does not correctly unwind from
the resources that was allocated before the call to
platform_driver_register().

As platform_driver_register() should never really fail here, I'll just
revert this patch and maybe someone, if they care about this driver, can
unwind the mess here.  Given that this driver doesn't look to ever be
able to be unloaded from the system, failing the init call does not look
to be a wise idea.

Cc: Aditya Pakki <pakki001@umn.edu>
Cc: Sinan Kaya <okaya@kernel.org>
Cc: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>