Revert "atl1e: checking the status of atl1e_write_phy_reg"
This reverts commit ff07d48d7bc0974d4f96a85a4df14564fb09f1ef.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

The original commit claimed to "fix" an issue here, but all that happens
is an error message is sent to the log, no actual recovery or error
handling happens at all.  Revert this and fix up the problem properly
with a follow-on change.

Cc: Kangjie Lu <kjlu@umn.edu>
Cc: David S. Miller <davem@davemloft.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>