Revert "hwmon: (lm80) fix a missing check of bus read in lm80 probe"
This reverts commit 9aa3aa15f4c2f74f47afd6c5db4b420fadf3f315.

Commits from @umn.edu addresses have been found to be submitted in "bad
faith" to try to test the kernel community's ability to review "known
malicious" changes.  The result of these submissions can be found in a
paper published at the 42nd IEEE Symposium on Security and Privacy
entitled, "Open Source Insecurity: Stealthily Introducing
Vulnerabilities via Hypocrite Commits" written by Qiushi Wu (University
of Minnesota) and Kangjie Lu (University of Minnesota).

This change was not needed, so revert it.  Also, the call to
lm80_init_client() is not properly handled, so if error handling is
needed in the lm80_probe() function, then it should be done properly,
not half-baked like the commit being reverted here did.

Cc: Kangjie Lu <kjlu@umn.edu>
Acked-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>